make wrappable in pipeline
