+++ pom.xml
@@ -3,7 +3,7 @@
     <parent>
         <groupId>org.jenkins-ci.plugins</groupId>
         <artifactId>plugin</artifactId>
-        <version>1.532.1</version> <!-- earliest version that requires Java 6 -->
+        <version>2.21</version>
     </parent>
 
     <artifactId>plain-timestamper</artifactId>
