+++ src/main/java/hudson/plugins/timestamper/TimestamperBuildWrapper.java
@@ -23,25 +23,24 @@
  */
 package hudson.plugins.timestamper;
 
-import static com.google.common.base.Preconditions.checkNotNull;
+import hudson.EnvVars;
 import hudson.Extension;
+import hudson.FilePath;
 import hudson.Launcher;
-import hudson.console.LineTransformationOutputStream;
-import hudson.model.BuildListener;
+import hudson.console.ConsoleLogFilter;
 import hudson.model.AbstractBuild;
 import hudson.model.AbstractProject;
 import hudson.model.Run;
+import hudson.model.TaskListener;
 import hudson.tasks.BuildWrapper;
 import hudson.tasks.BuildWrapperDescriptor;
+import jenkins.tasks.SimpleBuildWrapper;
+import org.kohsuke.stapler.DataBoundConstructor;
 
 import java.io.IOException;
 import java.io.OutputStream;
-import java.io.OutputStreamWriter;
-import java.util.Date;
+import java.io.Serializable;
 import java.util.logging.Logger;
-
-import org.apache.commons.lang.time.FastDateFormat;
-import org.kohsuke.stapler.DataBoundConstructor;
 
 /**
  * Build wrapper that decorates the build's logger to record time-stamps as each
@@ -51,9 +50,8 @@ import org.kohsuke.stapler.DataBoundConstructor;
  *
  * @author Steven G. Brown
  */
-public final class TimestamperBuildWrapper extends BuildWrapper {
+public final class TimestamperBuildWrapper extends SimpleBuildWrapper implements Serializable{
 
-    private static final String DEFAULT_TIME_FORMAT = "HH:mm:ss";
     private static final Logger LOGGER = Logger.getLogger(TimestamperBuildWrapper.class.getName());
 
     /**
@@ -66,6 +64,24 @@ public final class TimestamperBuildWrapper extends BuildWrapper {
     /**
      * {@inheritDoc}
      */
+    @Override
+    public void setUp(Context context, Run<?, ?> build, FilePath workspace, Launcher launcher,
+                      TaskListener listener, EnvVars initialEnvironment) throws IOException, InterruptedException {
+        // nothing to do
+    }
+
+
+    /**
+     * {@inheritDoc}
+     */
+    @Override
+    public ConsoleLogFilter createLoggerDecorator(Run<?, ?> build) {
+        return new ConsoleLogFilterImpl(build);
+    }
+
+/*    *//**
+     * {@inheritDoc}
+     *//*
     @SuppressWarnings("rawtypes")
     @Override
     public Environment setUp(AbstractBuild build, Launcher launcher, BuildListener listener)
@@ -74,73 +90,7 @@ public final class TimestamperBuildWrapper extends BuildWrapper {
         return new Environment() {
             // No tear down behaviour or additional environment variables.
         };
-    }
-
-    /**
-     * {@inheritDoc}
-     */
-    @SuppressWarnings("rawtypes")
-    @Override
-    public OutputStream decorateLogger(AbstractBuild build, OutputStream logger) {
-        return new TimestamperOutputStream(logger, build);
-    }
-
-
-    /**
-     * Output stream that writes each line to the provided delegate output stream
-     * after inserting a {@link TimestampNote}.
-     */
-    private static class TimestamperOutputStream extends LineTransformationOutputStream {
-
-        /**
-         * The delegate output stream.
-         */
-        private final OutputStream delegate;
-        private Run<?, ?> build;
-        private OutputStreamWriter streamWriter;
-
-        /**
-         * Create a new {@link TimestampNotesOutputStream}.
-         *
-         * @param delegate
-         *            the delegate output stream
-         */
-        TimestamperOutputStream(OutputStream delegate, Run<?, ?> build) {
-            this.delegate = checkNotNull(delegate);
-            this.streamWriter = new OutputStreamWriter(delegate);
-            this.build = build;
-        }
-
-        private String format(Timestamp timestamp) {
-            // Jenkins.
-            // TODO get global configuration which defines timeFormat then
-            String template = "[%s] ";
-            String systemTime = FastDateFormat.getInstance(DEFAULT_TIME_FORMAT).format(new Date(timestamp.millisSinceEpoch));
-            return String.format(template, systemTime);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        protected void eol(byte[] b, int len) throws IOException {
-            OutputStreamWriter streamWriter = new OutputStreamWriter(delegate);
-            Timestamp timestamp = new Timestamp(System.currentTimeMillis() - build.getTimeInMillis(), System
-                .currentTimeMillis());
-            streamWriter.write(format(timestamp));
-            streamWriter.flush();
-            delegate.write(b, 0, len);
-        }
-
-        /**
-         * {@inheritDoc}
-         */
-        @Override
-        public void close() throws IOException {
-            super.close();
-            delegate.close();
-        }
-    }
+    }*/
 
     /**
      * Registers {@link TimestamperBuildWrapper} as a {@link BuildWrapper}.
@@ -163,6 +113,16 @@ public final class TimestamperBuildWrapper extends BuildWrapper {
         @Override
         public boolean isApplicable(AbstractProject<?, ?> item) {
             return true;
+        }
+    }
+
+    private class ConsoleLogFilterImpl extends ConsoleLogFilter implements Serializable {
+        public ConsoleLogFilterImpl(Run<?, ?> build) {
+        }
+
+        @Override
+        public OutputStream decorateLogger(AbstractBuild abstractBuild, OutputStream logger) throws IOException, InterruptedException {
+            return new TimestamperOutputStream(logger, abstractBuild);
         }
     }
 }
